requirements.txt
appdirs==1.4.3
imgdiff==1.7.0
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.2.0
selenium==3.3.0
six==1.10.0
